Major Update to Champions Page
This update incorporates the changes required by the python backend
(accesses different database collections). The UI was also updated to
include items with highest winrates (with a brief explanation regarding
our thought process). Template "About" page was brought back and will be
completed later to explain how the app works. Also created dynamic links
to data dragon so resources introduced in later patches (i.e. new item
icons and champion images) can be retrieved without having to manually
update the data dragon source URL.

Summoners page still has to update the data dragon URLs and account for
the deprecation of the MatchHistory API call. Those pages and
controllers were left untouched so the changes could be made uniformly.

The major thing in this update was the back-end script, which is now
intelligent enough to determine the patch a match was played on, and
will only generate smart item sets for the current patch. Most of the
changes in this commit relate to getting that working again.